Training with EnKF
Here, the training of neural network with ensemble filter is tested
(much lower cost than UKF or CKF). The accuracy achieved is about ~90%
which is lower than that of SGD (Stochastic Gradient Descent). This is
to be expected since a mere 150 ensemble members was used and the
obvservation space is expected to be spanned by only this dimensionalty.
The test was conducted on the MINIST data set itself rather than the on
regression (non-linear functions).